fs: Fix RenameFile for directory renames
When renaming a file that ends up to be a directory in the auxiliary
directory we need to make sure that all ZenFS children files are also
renamed.

ZenFS does not implement the concept of directories and rather creates
files which have the absolute path (relative from the auxiliary
directory) as a filename. Thus we need to rename all ZenFS files that
relate to an auxiliary directory that is being renamed.

Signed-off-by: Dennis Maisenbacher <dennis.maisenbacher@wdc.com>